Experimental cross-compiling CMake toolchain
Invoke with

    mkdir xbuild
    cd xbuild
    cmake -DCMAKE_TOOLCHAIN_FILE=../cmake_scripts/Toolchain-mingw32.cmake ..

Currently has the bug that the texbake and fontbake tools also get built with
the cross-compiler, and are then unrunnable. This can be patched by manually
copying over the tool binaries, and then re-running make. Beyond that, some
random tweaks are needed before an executable can be made, CMake's find_package
for Boost doesn't work with the mingw environment, and the resulting executable
didn't run on Wine. More work is needed before cross-compiling is viable.