Tile work
Had an idea to use a different wall tile for the indoor sides of
walls, but to implement that properly I'd also need to do two sets
of the doorway and window tiles. Too much work for a visual effect
at this point.

Saved the indoor wall tiles anyway.

For even more complexity, make the indoor wall have a different
color than the outdoor one, and rig the wall drawing system to use
multiple colors so that the top and side strips are always in
outdoor wall colors. A multipart wall drawing scheme might not be
entirely stupid, since it might let us get away with fewer sprites
per wall type and clever combinations of them.