Refactor worldview
Ditched the Location space / chart dichotomy that's only needed for
portaling spaces.

Draw fog of war on a higher z-layer (this changes visuals a bit).
The idea is that now the FX layer under the fog layer can be used to
draw effects like beams and explosions that get covered by the FOV
shroud.