ECS and world architecture refactoring start
I want to use the new librarized entity component system, but that
means abandoning the current entity datatype and the call API. And
getting rid of the singleton World variable, that was a bit iffy to
begin with, since the new Entity type isn't that good for a lot of
custom methods.

I'll just start breaking up everything then.