Fix item layering, show items in map memory
Items need to show on top of the content in the cell behind them. Items
are now drawn in the Object sort layer, and they are placed under mobs
by controlling the sprite insertion order.

Items now show up in map memory, though this is currently implemented
in the wrong way that uses live item data instead of saving a separate
item icon in the map memory data.