Preliminary script for building release binaries
The script pulls all the stops for minimizing binary size. It looks like
you can't easily make different configurations in the standard cargo
file, we want the much slower lto and single thread compile version for
the "actually a release" builds, but not for the "just make it fast"
regular --release ones.

Should do more stuff like zipping the binary in something that has the
version string in it, but this is a good starting point.