feat: Remove the FailingIncremental type
The old idea was that World could fail to update on a command that makes
no sense. This isn't really realistic though, as with the Incremental
pattern you'd need to either clone the world every frame or replay it
back from the beginning to recover from the error. (Rewriting it to have
a mutable self method for update would require the mutable-method to
have an unenforceable guarantee to change nothing if the update fails,
also iffy.)

New plan is that any event will actually update the world, and events
that the world can't make sense of will just reduce to no-ops that still
advance the clock.